--- os_in_browser/src/hooks/useFiles.ts
@@ -1,6 +1,0 @@
-//Mapriles from ProcessComsumer 
-
-export const useFiles = (directory:string ) =>{
-    return 
-};
-
